Remove final solution from exercise files
Some of the exercise files was the same (or similar) to the final
answer key. These have been “cleared out”, so they now only contain a
few helpful comments.

Note that some of them are not referenced as exercises in the video
material. There might also be some files still, that are meant to be
exercises, but which contain the solution.